% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/site-academic.R
\name{create_site_academic}
\alias{create_site_academic}
\title{Create a hugo academic site}
\usage{
create_site_academic(
  path = ".",
  open = is_interactive(),
  rstudio = rstudioapi::isAvailable()
)
}
\arguments{
\item{path}{Path to create site}

\item{open}{Open new site after creation?}

\item{rstudio}{Create RStudio project?}
}
\description{
Create a hugo academic 4.8.0 site, configured to work well with hugodown.
In particular, it ensures that the following features important for R
users work correctly:
\itemize{
\item Syntax highlighting (turns off default js highlighting, renables
default chroma, and sets up styles in \code{assets/chroma.css}).
\item Math
\item HTML widgets
\item Default post archetype is tweaked to create \code{.Rmd}
}
}
