% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/site-calade.R
\name{create_site_calade}
\alias{create_site_calade}
\title{Create a hugo calade site}
\usage{
create_site_calade(
  path = ".",
  knit = TRUE,
  open = rlang::is_interactive(),
  rstudio = rstudioapi::isAvailable()
)
}
\arguments{
\item{path}{Path to create site}

\item{knit}{Should hugodown attempt to knit R markdown files?}

\item{open}{Open new site after creation?}

\item{rstudio}{Create RStudio project?}
}
\description{
Create a hugo calade site
}
