% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/site-vanilla.R
\name{create_site_vanilla}
\alias{create_site_vanilla}
\title{Create a hugo vanilla site}
\usage{
create_site_vanilla(
  path = ".",
  open = is_interactive(),
  rstudio = rstudioapi::isAvailable()
)
}
\arguments{
\item{path}{Path to create site}

\item{open}{Open new site after creation?}

\item{rstudio}{Create RStudio project?}
}
\description{
Create a \href{https://github.com/zwbetz-gh/vanilla-bootstrap-hugo-theme}{hugo vanilla site},
configured to work well with hugodown.
In particular, it ensures that the following features important for R
users work correctly:
\itemize{
\item Syntax highlighting (enables
default chroma, and sets up styles in \code{assets/chroma.css}).
\item Math
\item HTML widgets
\item Default post archetype is tweaked to create \code{.Rmd}
}
}
