% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortcode.R
\name{embed_gist}
\alias{embed_gist}
\alias{embed_instagram}
\alias{embed_tweet}
\alias{embed_vimeo}
\alias{embed_youtube}
\title{Generate hugo shortcodes to embed various types of media}
\usage{
embed_gist(username, id, filename = NULL)

embed_instagram(id, caption = TRUE)

embed_tweet(id, username = NULL)

embed_vimeo(id)

embed_youtube(id, autoplay = FALSE)
}
\arguments{
\item{username}{GitHub/Twitter user name}

\item{id}{A string giving the object id. You'll usually find this by
inspecting the URL:
\itemize{
\item gist: \verb{https://gist.github.com/spf13/7896402} -> \code{7896402}
\item instagram: \verb{https://www.instagram.com/p/BWNjjyYFxVx/} -> \code{BWNjjyYFxVx}
\item twitter: \verb{https://twitter.com/spf13/status/877500564405444608} -> \code{877500564405444608}
\item vimeo: \verb{https://vimeo.com/channels/staffpicks/146022717} -> \code{146022717}
\item youtube: \verb{https://www.youtube.com/watch?v=w7Ft2ymGmfc} -> \code{w7Ft2ymGmfc}
}}

\item{filename}{Pick single file from multiple file gist}

\item{caption}{Show instagram caption?}

\item{autoplay}{Automatically play youtube video?}
}
\description{
These are wrappers that make it easy to generate
\href{https://gohugo.io/content-management/shortcodes/}{hugo shortcodes} that
make it easy to embed various types of media into your pages. You use from
inline R code like:

\if{html}{\out{<div class="sourceCode">}}\preformatted{This tweet announced the release of hugo 0.24:

`R embed_tweet("877500564405444608")`
}\if{html}{\out{</div>}}
}
