% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md-document.R
\name{hugo_document}
\alias{hugo_document}
\title{Deprecated: please use \code{md_document()}}
\usage{
hugo_document(
  toc = FALSE,
  toc_depth = 3,
  fig_width = 7,
  fig_asp = 0.618,
  fig_retina = 2,
  tidyverse_style = TRUE,
  standalone = FALSE,
  includes = NULL,
  pandoc_args = NULL
)
}
\description{
Deprecated: please use \code{md_document()}
}
\keyword{internal}
