% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hugo-install.R
\name{hugo_install}
\alias{hugo_install}
\title{Install specified version of hugo}
\usage{
hugo_install(version = NULL, os = hugo_os(), arch = "64bit", extended = TRUE)
}
\arguments{
\item{version}{String giving version (e.g. "0.69.0"). If omitted will
default to latest release.}

\item{os}{Operating system, one of "Linux", "Windows", "macOS". Defaults
to current operating system.}

\item{arch}{Architecture}

\item{extended}{Installed hugo-extended which also includes SCSS etc?}
}
\description{
Downloads binary from hugo releases, and installs in system wide cache.
}
\examples{
\dontrun{
hugo_install()
}
}
