% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hugo.R
\name{hugo_locate}
\alias{hugo_locate}
\title{Locates a specific Hugo installation}
\usage{
hugo_locate(version = hugo_default_get())
}
\arguments{
\item{version}{The Hugo version to be located}
}
\value{
Path to the Hugo install
}
\description{
Locates a specific Hugo installation
}
