% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hugo-version.R
\name{hugo_version}
\alias{hugo_version}
\title{Find hugo version needed for current site}
\usage{
hugo_version(site = ".")
}
\arguments{
\item{site}{Path to hugo site.}
}
\description{
Hugo changes rapidly, so it's important to pin your site to a specific
version and then deliberately update when needed. This function reports
which of hugo your site correctly uses.

The primary location of this information is the \code{hugo_version} field
in \code{hugodown.yaml}. If that doesn't exist, we also look in \code{netlify.toml}.
}
\seealso{
\code{\link[=hugo_install]{hugo_install()}} to install any version of hugo.
}
