% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hugodown-package.R
\docType{package}
\name{hugodown-package}
\alias{hugodown}
\alias{hugodown-package}
\title{hugodown: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Hadley Wickham \email{hadley@rstudio.com}

Other contributors:
\itemize{
  \item RStudio [copyright holder]
}

}
\keyword{internal}
