% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md-document.R
\name{md_document}
\alias{md_document}
\title{An Rmd output format that produces Hugo-flavoured markdown}
\usage{
md_document(
  fig_width = 7,
  fig_asp = 0.618,
  fig_retina = 2,
  tidyverse_style = TRUE
)
}
\arguments{
\item{fig_width}{Figure width (in inches).}

\item{fig_asp}{Figure aspect ratio, defaults to the golden ratio.}

\item{fig_retina}{Scaling to perform for retina displays. Defaults to
\code{NULL} which performs no scaling. A setting of 2 will work for all
widely used retina displays, but will also result in the output of
\verb{<img>} tags rather than markdown images due to the need to set the
width of the image explicitly.}

\item{tidyverse_style}{Use tidyverse knitr conventions? This sets
\code{collapse = TRUE}, \verb{comment = "#>}, \code{fig.align = "center"}, and
\code{out.width = "700px"}.}
}
\description{
This RMarkdown output format is designed to generate markdown that is
maximally compatible with Hugo. It intelligently generates a preview so
that you see something useful when Hugo isn't running, but it doesn't
get in the way of hugo's full-site preview when it is.
}
\section{Syntax highlighting}{


\code{md_document()} uses a hybrid system for syntax highlighting.
For R code it uses \href{http://github.com/r-lib/downlit}{downlit}. For
other languages, it relies on Chroma, the syntax highlighter built into
hugo.
}

