% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post.R
\name{post_create}
\alias{post_create}
\title{Create a new post}
\usage{
post_create(path, kind = NULL, data = list(), site = ".")
}
\arguments{
\item{path}{Directory to create, like \code{blog/2020-my-favourite-package}.}

\item{kind}{Kind of archetype of use; usually automatically derived
from the base directory of \code{path}.}

\item{data}{Any additional data to be used when templating \code{.Rmd} files.
The default set \code{date} to today's date (in YYYY-MM-DD format), and
author to your name (if set in the \code{usethis.full_name} option).}

\item{site}{Path to the hugo site.}
}
\description{
Post creation takes advantage of Hugo's
\href{https://gohugo.io/content-management/archetypes/}{archetypes} or templates,
with an extension for \code{.Rmd} files. \code{post_create()} first calls \verb{hugo new}
(which will apply go templating to \code{.md} files in the archetype),
and then uses \href{https://github.com/edwindj/whisker}{whisker} to template
any \code{.Rmd} files.
}
