% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortcode.R
\name{shortcode}
\alias{shortcode}
\title{Generate a hugo shortcode}
\usage{
shortcode(
  .name,
  ...,
  .contents = NULL,
  .output = c("html", "md"),
  .inline = FALSE
)
}
\arguments{
\item{.name}{Name of the shortcode}

\item{...}{Arguments to the shortcode, supplied either by name or
position depending on the shortcode. By default, strings will
automatically be quoted with single quotes. Suppress this quoting
by wrapping the argument in \code{I()}.}

\item{.contents}{Contents of the shortcode for paired shortcodes.}

\item{.output}{Is the output of the shortcode html or markdown? This
controls whether the shortcode uses \verb{<>} or \verb{\%}.}

\item{.inline}{Is the shortcode designed to be used inline or in its own
paragraph? Controls whether the shortcode is wrapped in a block or inline
\href{https://pandoc.org/MANUAL.html#extension-raw_attribute}{raw attribute}.}
}
\description{
Generate a hugo shortcode with appropriate pandoc markup to preserve it as
is when embedded in an R markdown document.

Generally, I don't recommend calling this function directly; instead
use it inside a function with the same name as the shortcode you want to
wrap. See \code{\link[=embed_gist]{embed_gist()}} and friends for examples
}
\examples{
pkg <- function(name) {
  shortcode("pkg", name, .inline = TRUE)
}
pkg("hugodown")
}
