% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/site.R
\name{site_outdated}
\alias{site_outdated}
\title{Find \code{.Rmd}s that need to be re-rendered.}
\usage{
site_outdated(site = ".")
}
\arguments{
\item{site}{Path to hugo site.}
}
\description{
\code{\link[=hugo_document]{hugo_document()}} adds a hash of the input \code{.Rmd} in the YAML metdata of
the \code{.md} file that it creates. This provides a reliable way to determine
whether or not a \code{.Rmd} has been changed since the last time the \code{.md}
was rendered.
}
