% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/site.R
\name{site_rmd}
\alias{site_rmd}
\title{List all \code{.Rmd}s in a hugo site}
\usage{
site_rmd(needs_render = TRUE, site = ".")
}
\arguments{
\item{needs_render}{If set to \code{TRUE}, will only list \code{.Rmd}s that
need to be updated. If \code{FALSE}, lists all \code{.Rmds}.}

\item{site}{Path to hugo site.}
}
\description{
\code{\link[=hugo_document]{hugo_document()}} adds a hash of the input \code{.Rmd} in the YAML metdata of
the \code{.md} file that it creates. This provides a reliable way to determine
}
