% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.R
\name{tidy_post_create}
\alias{tidy_post_create}
\alias{tidy_thumbnails}
\alias{tidy_show_meta}
\title{Various helpers for tidyverse.org and similar sites}
\usage{
tidy_post_create(slug, site = ".")

tidy_thumbnails(path = NULL)

tidy_show_meta(min = 1, site = ".")
}
\arguments{
\item{slug}{File name of new post. Year and month will be automatically
appended.}

\item{site}{Path to hugo site}

\item{path}{Path to blog post}

\item{min}{Minimum number of uses}
}
\description{
\itemize{
\item \code{tidy_post_create()} makes a new post
\item \code{tidy_thumbnails()} resizes thumbnails to the correct size
\item \code{tidy_show_meta()} prints tags and categories used by existing posts.
}
}
