% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netlify.R
\name{use_netlify_toml}
\alias{use_netlify_toml}
\title{Create \code{netlify.toml}}
\usage{
use_netlify_toml(site = ".")
}
\arguments{
\item{site}{Path to the hugo site.}
}
\description{
This helper creates a basic \code{netlify.toml} file, automatically setting the
hugo version to match your blog. This is needed when publishing your site
with \href{https://www.netlify.com}{netlify}; see \code{vignette("deploy")} for more
details.
}
