% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post.R
\name{use_post}
\alias{use_post}
\title{Create a new post}
\usage{
use_post(path, kind = NULL, data = list(), site = ".", open = is_interactive())
}
\arguments{
\item{path}{Directory to create, like \code{blog/2020-my-favourite-package}.}

\item{kind}{Kind of archetype of use; usually automatically derived
from the base directory of \code{path}.}

\item{data}{Any additional data to be used when templating \code{.Rmd} files.

The default data includes:
\itemize{
\item \code{date}: today's date (in YYYY-MM-DD format).
\item \code{author}: \code{\link[whoami:fullname]{whoami::fullname()}}.
\item \code{slug}: taken from the file name of \code{path}.
}}

\item{site}{Path to the hugo site.}

\item{open}{Open file for interactive editing?}
}
\description{
Post creation takes advantage of Hugo's
\href{https://gohugo.io/content-management/archetypes/}{archetypes} or templates,
with an extension for \code{.Rmd} files. \code{use_post()} first calls \verb{hugo new}
(which will apply go templating to \code{.md} files in the archetype),
and then uses \href{https://github.com/edwindj/whisker}{whisker} to template
any \code{.Rmd} files.
}
