% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.R
\name{use_tidy_post}
\alias{use_tidy_post}
\alias{use_tidy_thumbnails}
\alias{tidy_show_meta}
\title{Various helpers for tidyverse.org and similar sites}
\usage{
use_tidy_post(slug, site = ".", open = is_interactive())

use_tidy_thumbnails(path = NULL)

tidy_show_meta(min = 1, site = ".")
}
\arguments{
\item{slug}{File name of new post. Year and month will be automatically
appended.}

\item{site}{Path to the hugo site.}

\item{open}{Open file for interactive editing?}

\item{path}{Path to blog post}

\item{min}{Minimum number of uses}
}
\description{
\itemize{
\item \code{use_tidy_post()} makes a new post
\item \code{use_tidy_thumbnails()} resizes thumbnails to the correct size
\item \code{tidy_show_meta()} prints tags and categories used by existing posts.
}
}
