% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hugo-server.R
\name{hugo_build}
\alias{hugo_build}
\title{Build site}
\usage{
hugo_build(site = ".", build_drafts = FALSE, build_future = FALSE)
}
\arguments{
\item{site}{Path to hugo site.}

\item{build_drafts, build_future}{Should drafts and future posts be included
in the built site?}
}
\description{
Build hugo site into \verb{public/} directory. Useful for debugging and some
deployment scenarios
}
