% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{hugo_document}
\alias{hugo_document}
\title{Convert to Hugo-flavoured markdown}
\usage{
hugo_document(fig_width = 7, fig_height = 5, fig_retina = NULL)
}
\arguments{
\item{fig_width, fig_height}{Default figure dimensions (in inches).}
}
\description{
This RMarkdown output format is designed to generate markdown that is
maximally compatible with Hugo. It intelligently generates a preview so
that you see something useful when Hugo isn't running, but it doesn't
get in the way of hugo's full-site preview when it is.
}
