% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hugo-format.R
\name{hugo_document}
\alias{hugo_document}
\title{Deprecated: please use \code{md_document()}}
\usage{
hugo_document(
  fig_width = 7,
  fig_asp = 0.618,
  fig_retina = NULL,
  tidyverse_style = TRUE
)
}
\description{
Deprecated: please use \code{md_document()}
}
\keyword{internal}
