% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hugo-format.R
\name{hugo_document}
\alias{hugo_document}
\title{Convert to Hugo-flavoured markdown}
\usage{
hugo_document(
  fig_width = 7,
  fig_asp = 0.618,
  fig_retina = NULL,
  tidyverse_style = TRUE,
  downlit = FALSE
)
}
\arguments{
\item{fig_width}{Figure width (in inches).}

\item{fig_asp}{Figure aspect ratio, defaults to the golden ratio.}

\item{fig_retina}{Scaling to perform for retina displays. Defaults to
\code{NULL} which performs no scaling. A setting of 2 will work for all
widely used retina displays, but will also result in the output of
\code{<img>} tags rather than markdown images due to the need to set the
width of the image explicitly.}

\item{tidyverse_style}{Use tidyverse knitr conventions? This sets
\code{collapse = TRUE}, \verb{comment = "#>}, \code{fig.align = "center"}, and
\code{out.width = "700px"}.}

\item{downlit}{If \code{TRUE}, uses \href{http://github.com/r-lib/downlit}{downlit}
for syntax highlighting. This converts code blocks to \verb{<pre>} blocks
and adds links around inline code. If \code{FALSE}, syntax highlighting
will be done by hugo.}
}
\description{
This RMarkdown output format is designed to generate markdown that is
maximally compatible with Hugo. It intelligently generates a preview so
that you see something useful when Hugo isn't running, but it doesn't
get in the way of hugo's full-site preview when it is.
}
