% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hugodown-package.R
\docType{package}
\name{hugodown-package}
\alias{hugodown}
\alias{hugodown-package}
\title{hugodown: Make websites with hugo and RMarkdown}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

hugodown is a minimal version of blogdown that only supports hugo, and more clearly delineates the roles of RMarkdown (turning .Rmd into .md) and hugo (turning .md into .html).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/r-lib/hugodown}
  \item Report bugs at \url{https://github.com/r-lib/hugodown/issues}
}

}
\author{
\strong{Maintainer}: Hadley Wickham \email{hadley@rstudio.com}

Other contributors:
\itemize{
  \item RStudio [copyright holder]
}

}
\keyword{internal}
