% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{server_start}
\alias{server_start}
\alias{server_stop}
\alias{server_browse}
\title{Manage the hugo server}
\usage{
server_start(
  site = ".",
  auto_navigate = TRUE,
  browse = TRUE,
  render_to_disk = FALSE
)

server_stop()

server_browse()
}
\arguments{
\item{site}{Path to hugo site.}

\item{auto_navigate}{Automatically navigate to the most recently changed
page?}

\item{browse}{Automatically preview the site after the server starts?}

\item{render_to_disk}{Render site to disk? The default is to serve the
site from memory, but rendering to disk can be helpful for debugging.}
}
\description{
\code{server_start()} starts a hugo server that will automatically re-generate
the site whenever the input changes. You only need to execute this once
per session; it continues to run in the background as you work on the site.

\code{server_stop()} kills the server. This happens automatically when you exit
R so you shouldn't normally need to run this.

\code{server_browse()} opens the site in the RStudio viewer or your web browser.
}
